ARM: dts: msm: disable memcg kernel and socket accounting on sdm670
Disable memcg kernel and socket accounting on sdm670 since memory
cgroups is used in this target as a means of user space task
grouping based on oom_score_adj, and controlling the userspace
memory consumed by tasks. Kernel and socket memory accounting is
not considered. Since memcg is created per process on this target
enabling kernel accounting results in the creation of numerous
kmem caches resulting in significant overhead.

Change-Id: I2caad9ce6cca5846b6183a0f0753977db54a660d
Signed-off-by: Vinayak Menon <vinmenon@codeaurora.org>

[ab123321] Apply sm8150 commit to sdm670
Signed-off-by: Artem Labazov <123321artyom@gmail.com>
Signed-off-by: ethan-halsall <ethan.halsall@gmail.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>
Change-Id: I9dea3f630320b6520568d669b919213069a18b37
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>