asoc: msm-pcm-routing: add msm_pcm_routing_get_pp_ch_cnt
Based on the type of pre/post processing, it is possible that
the processed channel count could be different from the input
channel count. As an example, the FFECNS processing block
could accept 3 channel input and has the ability to output
upto 9 channels of processed data. In cases like these, the
stream side needs to know the processed channel count to set
up its own configuration. Change adds functionality in
routing driver to read the processed channel count

Change-Id: I06c365f8c53270712709631bc23649d37434f405
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: Avinash Chandra <avicha@codeaurora.org>