lsm_client: add support to perform channel mixing
With pre-processing applied, it is possible that the processed
number of channels could be different than the number of channels
that voice wakeup can run with. In such cases, the channel mixer
should be setup to downmix the channel. For voice wakeup to detect
keywords properly, preprocessed channels cannot be mixed hence
the channel mixer is always setup with unity co-efficients or 0
to either select or drop the channel. In this case the channel mixer
acts more like an channel selector

Change-Id: I79b578b56ad5484e27f433404a5371540618a2ae
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: Avinash Chandra <avicha@codeaurora.org>