msm: thermal: simple: Introduce simple MSM thermal solution
When properly configured, this driver monitors temperatures from a
specified sensor and reduces the CPU's max frequency in response to
the temperature crossing specified thermal zones. The temperature sensor
use must be a specified ADC channel on a given VADC device, hence this
driver's dependency on the chipset being an MSM product.

This driver can only be configured via the device tree; it cannot be
configured at runtime. Configuration instructions can be found in the
accompanying documentation.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>