qos: Remove pm_qos_update_request_timeout() API
Using a timeout for a PM QoS request can lead to disastrous results on
power consumption. It's always possible to find a fixed scope in which a
PM QoS request should be applied, so timeouts aren't ever strictly
needed; they're usually just a lazy way of using PM QoS. Remove the API
so that it cannot be abused any longer.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[KangMamles: Port to 4.9]
Signed-off-by: KangMamles <imzea69@gmail.com>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>