UPSTREAM: linux/jiffies.h: Extract common header for vDSO
The vDSO library should only include the necessary headers required for
a userspace library (UAPI and a minimal set of kernel headers). To make
this possible it is necessary to isolate from the kernel headers the
common parts that are strictly necessary to build the library.

Split jiffies.h into linux and common headers to make the latter suitable
for inclusion in the vDSO library.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@arm.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lkml.kernel.org/r/20200320145351.32292-14-vincenzo.frascino@arm.com
(cherry picked from commit 97b01d2eabd8c0fa650a9d513febeae35dbc6de4)
Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 154668398
Change-Id: I68b8450986462838343cb3e1badcf58a5f739b26
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>