treewide: Remove is_project() and its usage
is_project() is only used in conditions to conditionally execute codes specific to a device.
The problem is, is_project()'s value would never change, Thus, it really doesn't make sense to use runtime conditions for it.

As, we only target Realme XT (OPPO_19651) & Realme 5 Pro (OPPO_19691),
Remove conditions for other phones and use compile time conditions that make use of kconfig values introduced in [1] for rset of the conditions

[1] https://github.com/dakkshesh07/parallax_kernel_realme_rmx1921/commit/79bab63a3376a22c7352f4eb2c3cb9bf3e9bb49d