drivers/nfc: pn553: Add check for NFC sku
Our device has 2 variants (EU and IN) and only EU variants have NFC hardware.

This commit makes use of Oplus NFC drivers APIs to detect if the device
supports NFC and exposes and sysfs prop with the results.
It also bails out of the init process if it doesn't detect NFC hardware.
The sysfs prop can be used by our rom side init scripts to disable NFC services for IN variants.

Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>